# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_interfaces_http_handlers.ConfirmInvitationRequest:
    properties:
      couple_id:
        type: string
    required:
    - couple_id
    type: object
  internal_interfaces_http_handlers.CoupleResponse:
    properties:
      couple_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      couple_name:
        type: string
      created_at:
        type: string
      invitation_expires_at:
        type: string
      relationship_start_date:
        type: string
      relationship_type:
        example: dating
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
      user1_id:
        type: string
      user2_id:
        type: string
    type: object
...
  internal_interfaces_http_handlers.DashboardResponse:
    properties:
      couple:
        properties:
          couple_id:
            type: string
          partner_name:
            type: string
          relationship_type:
            type: string
        type: object
      last_updated:
        type: string
      mood_history:
        items:
          properties:
            date:
              type: string
            intensity:
              type: integer
            mood_type:
              type: string
          type: object
        type: array
      streaks:
        items:
          properties:
            activity_type:
              type: string
            current_streak:
              type: integer
            longest_streak:
              type: integer
            status:
              type: string
          type: object
        type: array
      todays_mood:
        properties:
          intensity:
            type: integer
          mood_type:
            type: string
        type: object
      user_id:
        type: string
    type: object
  internal_interfaces_http_handlers.ErrorResponse:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: validation failed
        type: string
    type: object
  internal_interfaces_http_handlers.HomeSummaryResponse:
    properties:
      active_streaks:
        type: integer
      couple_id:
        type: string
      longest_streak:
        type: integer
      partner_name:
        type: string
      relationship_type:
        type: string
      todays_mood_intensity:
        type: integer
      todays_mood_type:
        type: string
    type: object
  internal_interfaces_http_handlers.InviteRequest:
    properties:
      couple_name:
        type: string
      invite_key:
        type: string
      relationship_type:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_shared.RelationshipType'
        enum:
        - dating
        - engaged
        - married
    required:
    - invite_key
    type: object
  internal_interfaces_http_handlers.InviteResponse:
    properties:
      couple_id:
        type: string
      invitation_expires_at:
        type: string
      message:
        type: string
      status:
        $ref: '#/definitions/loviary_app_backend_internal_domain_shared.CoupleStatus'
    type: object
...
  internal_interfaces_http_handlers.MessageResponse:
    properties:
      message:
        example: Operation completed
        type: string
    type: object
...
  loviary_app_backend_internal_domain_shared.CoupleStatus:
    enum:
    - pending_invitation
    - active
    - grace_period
    - ended
    type: string
    x-enum-varnames:
    - CoupleStatusPendingInvitation
    - CoupleStatusActive
    - CoupleStatusGracePeriod
    - CoupleStatusEnded
...
  loviary_app_backend_internal_domain_shared.RelationshipType:
    enum:
    - dating
    - engaged
    - married
    type: string
    x-enum-varnames:
    - RelationshipDating
    - RelationshipEngaged
    - RelationshipMarried
...
host: localhost:8080
info:
  contact:
    email: support@loviary.app
    name: API Support
  description: Loviary - Couple Relationship Tracking API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Loviary API
  version: "1.0"
paths:
...
  /couples/confirm:
    post:
      consumes:
      - application/json
      description: Accept a couple invitation sent to the user
      parameters:
      - description: Confirmation request with couple ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.ConfirmInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.MessageResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Couple not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm couple invitation
      tags:
      - couples
  /couples/invite:
    post:
      consumes:
      - application/json
      description: Send a couple invitation to another user using their invite key
      parameters:
      - description: Invite request with partner's invite key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.InviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Invitation created
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.InviteResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: User with invite key not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "409":
          description: Invitation conflict
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create couple invitation
      tags:
      - couples
  /couples/me:
    delete:
      consumes:
      - application/json
      description: Initiate breakup of current couple relationship (grace period applies)
      produces:
      - application/json
      responses:
        "200":
          description: Breakup initiated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.MessageResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: No active couple found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete my couple
      tags:
      - couples
    get:
      consumes:
      - application/json
      description: Get current couple relationship information
      produces:
      - application/json
      responses:
        "200":
          description: Couple data
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.CoupleResponse'
        "404":
          description: No active couple found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my couple
      tags:
      - couples
  /home:
    get:
      consumes:
      - application/json
      description: Get aggregated dashboard data including couple info, mood, streaks
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard data
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.DashboardResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard
      tags:
      - home
  /home/summary:
    get:
      consumes:
      - application/json
      description: Get simplified summary for home widget display
      produces:
      - application/json
      responses:
        "200":
          description: Summary data
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.HomeSummaryResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get home summary
      tags:
      - home
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"